set pregnant action
