+++ client/src/redux/features/animals/animalsThunks.js
@@ -1,7 +1,7 @@
 import axios from "axios";
 import { header } from "../../../constants/token";
 import { URL, URL_GET_STATS, URL_GET_TYPES_OF_ANIMALS, URL_POST_ANIMAL, URL_SEARCH_QUERY } from "../../../constants/urls";
-import { newAnimal, searchedAnimal, setAllAnimals, setDeleted, setStats, setTypeOfAnimals, update } from "./animalsSlice";
+import { newAnimal, searchedAnimal, setAllAnimals, setDeleted, setPregnant, setStats, setTypeOfAnimals, update } from "./animalsSlice";
 
 export const createNewAnimal = (obj, token) => {
   return async function (dispatch) {
@@ -89,6 +89,20 @@ export function getStats(token) {
       return dispatch(setStats(response.data));
     } catch (error) {
       return dispatch(setStats({ error: error.message }));
+    }
+  };
+}
+
+export function getPregnantAsc(token) {
+  return async function (dispatch) {
+    try {
+      const response = await axios.get(
+        URL + "animal/isPregnant?status=true&order=ASC",
+        header(token)
+      );
+      return dispatch(setPregnant(response.data));
+    } catch (error) {
+      return dispatch(setPregnant({ error: error.message }));
     }
   };
 }
